% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FADvisor-package.R
\docType{package}
\name{FADvisor-package}
\alias{FADvisor}
\alias{FADvisor-package}
\title{FADvisor: FADvisor}
\description{
This R package provides a suite of functions designed to generate key indicators for Fish Aggregating Devices (FADs). It simplifies the process of analyzing FAD-related data, offering insights into their usage, environmental impact, and overall performance. Ideal for researchers and environmentalists focused on sustainable fishing practices.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TahaImz/FADvisor}
  \item Report bugs at \url{https://github.com/TahaImz/FADvisor/issues}
}

}
\author{
\strong{Maintainer}: Taha Imzilen \email{imzilen.taha@gmail.fr} (\href{https://orcid.org/0000-0001-8257-9113}{ORCID})

}
\keyword{internal}
